# openapi.yaml
openapi: 3.0.0
info:
  title: E-Learning API Documentation
  description: |
    Dokumentasi RESTful API untuk Sistem E-Learning berbasis Laravel 12.
    
    ### Fitur Utama:
    * **Autentikasi:** Laravel Sanctum (Token-Based).
    * **RBAC (Role-Based Access Control):** admin, guru, siswa.
    * **Manajemen Akademik:** Kelas & Mata Pelajaran.
    * **Materi Pembelajaran:** Upload berkas (PDF/Gambar) dan Tautan Video.
    * **Forum Diskusi:** Pembuatan topik, komentar, dan balasan komentar bertingkat.
    * **Pengumuman:** Pengumuman sekolah & pengumuman spesifik per kelas.
    * **Notifikasi:** Notifikasi real-time untuk topik baru, komentar baru, dan balasan komentar.
  version: 1.0.0
servers:
  - url: http://localhost:8000/api
    description: Development Server Lokal
security:
  - BearerAuth: []
paths:
...
  /kelas/{id}:
    put:
      summary: Update Kelas (Admin, Guru Only)
      description: Memperbarui informasi data kelas.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: ID Kelas
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nama_kelas:
                  type: string
                  example: X MIPA 1 (Unggulan)
                tingkat:
                  type: string
                  example: X
                jurusan:
                  type: string
                  example: MIPA
                guru_id:
                  type: integer
                  example: 2
                tahun_ajaran:
                  type: integer
                  example: 2024
      responses:
        '200':
          description: Kelas berhasil diperbarui.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Kelas berhasil diperbarui.
                  data:
                    $ref: '#/components/schemas/Kelas'
        '401':
          $ref: '#/components/responses/401Unauthenticated'
        '403':
          $ref: '#/components/responses/403Forbidden'
        '404':
          $ref: '#/components/responses/404NotFound'
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'

    delete:
      summary: Hapus Kelas (Admin, Guru Only)
      description: Menghapus kelas secara permanen beserta data relasi siswanya.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: ID Kelas
      responses:
        '200':
          description: Kelas berhasil dihapus.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Kelas berhasil dihapus.
        '401':
          $ref: '#/components/responses/401Unauthenticated'
        '403':
          $ref: '#/components/responses/403Forbidden'
        '404':
          $ref: '#/components/responses/404NotFound'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Masukkan access token (plainTextToken) yang didapatkan saat login sukses.
  responses:
    401Unauthenticated:
      description: Pengguna belum login atau token tidak valid.
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: Unauthenticated. Silakan login terlebih dahulu.
    403Forbidden:
      description: Hak akses ditolak karena role tidak sesuai.
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: Akses ditolak. Anda tidak memiliki izin untuk mengakses resource ini.
              required_roles:
                type: array
                items:
                  type: string
                example: ["admin"]
              your_role:
                type: string
                example: siswa
    404NotFound:
      description: Resource / Data tidak ditemukan.
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: Data tidak ditemukan.
  schemas:
...
    Kelas:
      type: object
      properties:
        id:
          type: integer
          example: 1
        nama_kelas:
          type: string
          example: X MIPA 1
        tingkat:
          type: string
          example: X
        jurusan:
          type: string
          example: MIPA
        tahun_ajaran:
          type: integer
          example: 2024
        guru:
          type: object
          nullable: true
          properties:
            id: { type: integer, example: 2 }
            nama: { type: string, example: Budi Santoso }
            role: { type: string, example: guru }
        siswa:
          type: array
          items:
            type: object
            properties:
              id: { type: integer, example: 3 }
              nama: { type: string, example: Siti Rahayu }
              role: { type: string, example: siswa }
              pivot:
                type: object
                properties:
                  kelas_id: { type: integer, example: 1 }
                  siswa_id: { type: integer, example: 3 }
                  tahun_ajaran: { type: integer, example: 2024 }
...
    ValidationError:
      type: object
      properties:
        message:
          type: string
          example: Data yang dikirim tidak valid.
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          example:
            email:
              - "Email wajib diisi."
              - "Format email tidak valid."
...